Remove GenericStoredPaymentComponentProvider.
